+++ tests/e2e/support/index.js
@@ -3,3 +3,5 @@ import './signInAdminUser';
 import './signedInAsAdminUser';
 import './deleteTestDepartment';
 import './createTestDepartment';
+import './createProduct';
+import './deleteProduct';